requirements.txt
Django==3.2.17
django-import-export==3.0.2
gunicorn==20.1.0
asgiref==3.6.0
crispy-bootstrap4==2022.1
defusedxml==0.7.1
diff-match-patch==20200713
django-3-jet==1.0.8
django-crispy-forms==2.0
django-jet==1.0.6
django-material==1.11.3
django-material-admin==1.8.6
django-redis==5.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
google-api-python-client==1.4.1
httplib2==0.21.0
MarkupPy==1.14
oauth2client==4.1.3
odfpy==1.4.1
openpyxl==3.1.1
Pillow==9.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tablib==3.3.0
uritemplate==4.1.1
xlrd==2.0.1
xlwt==1.3.0
mysqlclient==2.1.1
